CHANGE: @W-17272495@: Refactor formatters into individual files
